requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Werkzeug==2.3.7
PyJWT==2.8.0
requests==2.31.0
